[miio] Fix duplicate MIoT actions and value ranges
Services of the same type, such as the main and side brush of a
vacuum, share their action names. The generated action values were
therefore identical, and as every action matching the selected value
is executed, resetting one consumable reset all of them. Duplicate
action values now get a counter suffix, like duplicate properties.

The value range was parsed as Integer, so uint32 and larger limits
overflowed to -1 and decimal limits and steps were truncated. It is
now parsed as BigDecimal.

The affected database files are corrected: the reset actions of
11 vacuum models get unique values and labels naming the consumable,
and the invalid maximum of -1 is removed from 9 read-only channels.

+++ bundles/org.openhab.binding.miio/src/main/java/org/openhab/binding/miio/internal/miot/PropertyDTO.java
@@ -12,6 +12,7 @@
  */
 package org.openhab.binding.miio.internal.miot;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import com.google.gson.annotations.Expose;
@@ -44,7 +45,7 @@ public class PropertyDTO {
     public List<OptionsValueDescriptionsListDTO> valueList = null;
     @SerializedName("value-range")
     @Expose
-    public List<Integer> valueRange = null;
+    public List<BigDecimal> valueRange = null;
     @SerializedName("unit")
     @Expose
     public String unit;
